+++ pyvortex/src/dataset.rs
@@ -12,9 +12,9 @@ use vortex::arrow::infer_schema;
 use vortex::dtype::FieldName;
 use vortex::error::VortexResult;
 use vortex::expr::{ident, ExprRef, Select};
-use vortex::file::v2::io::file::FileIoDriver;
-use vortex::file::v2::{Scan, VortexFile, VortexOpenOptions};
-use vortex::file::VortexRecordBatchReader;
+use vortex::file::io::file::FileIoDriver;
+use vortex::file::read::VortexRecordBatchReader;
+use vortex::file::{Scan, VortexFile, VortexOpenOptions};
 use vortex::io::{ObjectStoreReadAt, TokioFile, VortexReadAt};
 use vortex::sampling_compressor::ALL_ENCODINGS_CONTEXT;
 use vortex::stream::ArrayStream;
