+++ vortex-datafusion/src/persistent/cache.rs
@@ -6,7 +6,7 @@ use object_store::path::Path;
 use object_store::{ObjectMeta, ObjectStore};
 use vortex_array::ContextRef;
 use vortex_error::{vortex_err, VortexError, VortexResult};
-use vortex_file::v2::{FileLayout, VortexOpenOptions};
+use vortex_file::{FileLayout, VortexOpenOptions};
 use vortex_io::ObjectStoreReadAt;
 
 #[derive(Debug, Clone)]
