+++ vortex-datafusion/src/persistent/format.rs
@@ -23,8 +23,7 @@ use vortex_array::stats::Stat;
 use vortex_array::ContextRef;
 use vortex_dtype::FieldPath;
 use vortex_error::{vortex_err, VortexExpect, VortexResult};
-use vortex_file::v2::VortexOpenOptions;
-use vortex_file::VORTEX_FILE_EXTENSION;
+use vortex_file::{VortexOpenOptions, VORTEX_FILE_EXTENSION};
 use vortex_io::ObjectStoreReadAt;
 
 use super::cache::FileLayoutCache;
