+++ vortex-datafusion/src/persistent/opener.rs
@@ -15,7 +15,7 @@ use vortex_error::VortexResult;
 use vortex_expr::datafusion::convert_expr_to_vortex;
 use vortex_expr::transform::simplify_typed::simplify_typed;
 use vortex_expr::{and, ident, lit, select, ExprRef};
-use vortex_file::v2::{ExecutionMode, Scan, VortexOpenOptions};
+use vortex_file::{ExecutionMode, Scan, VortexOpenOptions};
 use vortex_io::ObjectStoreReadAt;
 
 use super::cache::FileLayoutCache;
