+++ vortex-file/src/exec/mod.rs
@@ -14,7 +14,7 @@ use vortex_error::VortexResult;
 /// pool.
 ///
 /// Note that the futures encapsulate heavy CPU code such as filtering and decompression. To
-/// offload keep I/O work separate, please see the [`crate::v2::io::IoDriver`] trait.
+/// offload keep I/O work separate, please see the [`crate::io::IoDriver`] trait.
 pub trait ExecDriver: Send + Sync {
     fn drive(
         &self,
