+++ vortex-file/src/exec/tokio.rs
@@ -5,7 +5,7 @@ use tokio::runtime::Handle;
 use vortex_array::ArrayData;
 use vortex_error::{vortex_err, VortexResult};
 
-use crate::v2::exec::ExecDriver;
+use crate::exec::ExecDriver;
 
 /// An [`ExecDriver`] implementation that spawns the futures onto a Tokio runtime.
 pub struct TokioDriver(pub Handle);
