+++ vortex-file/src/footer/segment.rs
