+++ vortex-file/src/segments/cache.rs
