+++ vortex-file/src/strategy.rs
