--- vortex-file/src/v2/mod.rs
@@ -1,15 +1,0 @@
-mod exec;
-mod file;
-mod footer;
-pub mod io;
-mod open;
-pub mod segments;
-mod strategy;
-#[cfg(test)]
-mod tests;
-mod writer;
-
-pub use file::*;
-pub use footer::FileLayout;
-pub use open::*;
-pub use writer::*;
