--- vortex-file/src/write/mod.rs
@@ -1,7 +1,0 @@
-pub use layout::LayoutSpec;
-pub use writer::VortexFileWriter;
-
-mod layout;
-mod postscript;
-mod stats_accumulator;
-mod writer;
